fix up object hooking and refactor
