+++ projects/pysan/pysan/pysan/sanlib.py
@@ -22,6 +22,8 @@ import time
 import os
 import subprocess
 
+from pysan import command_injection, redos
+
 sanitizer_log_level = 0
 def sanitizer_log(msg, log_level):
     global sanitizer_log_level
@@ -125,131 +127,23 @@ def pysan_add_hook(function: Callable[[Any], Any],
         if post_exec_hook is not None:
             tmp_ret = post_exec_hook(ret, *args, **kwargs)
             if tmp_ret is not None:
-                print("Overwriting ret value")
+                #print("Overwriting ret value")
                 ret = tmp_ret
         sanitizer_log(f"Hook end {str(function)}", 0)
         return ret
     return run
 
 
-##############
-# Sanitizers #
-##############
-
-#############################
-# Code injection sanitizers #
-#############################
-def check_code_injection_match(elem) -> Optional[str]:
-    # Check exact match
-    if elem == "exec-sanitizer":
-        return "Explicit command injection found."
-
-    # Check potential for injecting into a string
-    if "FROMFUZZ" in elem:
-        return "Fuzzer controlled content in data. Code injection potential."
-    return None
-
-
-def pysan_hook_subprocess_Popen(cmd, **kwargs):
-    """Hook for subprocess.Popen"""
-    # Check first argument
-    if type(cmd) is str:
-        res = check_code_injection_match(cmd)
-        if res != None:
-            raise Exception(
-                    f"Potental code injection in subprocess.Popen\n{res}")
-    if type(cmd) is list:
-        for elem in cmd:
-            res = check_code_injection_match(elem)
-            if res != None:
-                print(res)
-                raise Exception(
-                    f"Potential code injection in subprocess.Popen\n{res}")
-
-
-def pysan_hook_os_system(cmd):
-    """Hook for os.system"""
-    res = check_code_injection_match(cmd)
-    if res != None:
-        raise Exception(f"Potential code injection by way of os.system\n{res}")
-
-
-def pysan_hook_eval(cmd):
-    """Hook for eval"""
-    res = check_code_injection_match(cmd)
-    if res != None:
-        raise Exception(f"Potential code injection by way of eval\n{res}")
-
-
-# Hooks for regular expressions.
-# Main problem is to identify ReDOS attemps. This is a non-trivial task
-# - https://arxiv.org/pdf/1701.04045.pdf
-# - https://dl.acm.org/doi/pdf/10.1145/3236024.3236027
-# and the current approach we use is simply check for extensive computing time.
-# In essence, this is more of a refinement of traditional timeout checker from
-# the fuzzer, however, that's the consequence of ReDOS attacks as well.
-#
-# Perhaps the smartest would be to use something like e.g.
-# https://github.com/doyensec/regexploit to scan the regex patterns.
-# Other heuristics without going too technical on identifying super-linear
-# regexes:
-# - check
-#   - if "taint" exists in re.compile(xx)
-# - check 
-#   - for backtracking possbility in PATTERN within re.comile(PATTERN)
-#   - and
-#   - "taint" in findall(XX) calls.
-def pysan_hook_re_pattern_findall_post(self, s):
-    global starttime
-    print("In post hook")
-    try:
-        endtime = time.time() - starttime
-        if endtime > 4:
-            print("param: %s"%(s))
-            raise Exception("Potential ReDOS attack")
-    except NameError:
-        #print("For some reason starttime is not set, which it should have")
-        sys.exit(1)
-        pass
-
-def pysan_hook_re_pattern_findall_pre(self, s):
-    global starttime
-    starttime = time.time()
-    #time.sleep(5)
-    #print("Pattern")
-    #print(self.pattern)
-
-def pysan_hook_post_re_compile(retval, pattern, flags=None):
-    """Hook for re.compile post execution to hook returned objects functions"""
-    sanitizer_log("Inside of post compile hook", 0)
-    wrapper_object = create_object_wrapper(methods = {
-            "findall" : (pysan_hook_re_pattern_findall_pre, pysan_hook_re_pattern_findall_post)
-        }
-    )
-    hooked_object = wrapper_object(retval)
-    return hooked_object
-
-
-def pysan_hook_re_compile(pattern, flags=None):
-    """Check if tainted input exists in pattern. If so, likely chance of making
-    ReDOS possible."""
-    sanitizer_log("Inside re compile hook", 0)
-
-
-############################################
-# Set up the hooks
-############################################
 def pysan_add_hooks():
+    """Sets up hooks"""
     os.system = pysan_add_hook(os.system,
-                               pre_exec_hook = pysan_hook_os_system)
-    re.compile = pysan_add_hook(
-        re.compile,
-        pre_exec_hook = pysan_hook_re_compile,
-        post_exec_hook = pysan_hook_post_re_compile
-    )
+                               pre_exec_hook = command_injection.pysan_hook_os_system)
     subprocess.Popen = pysan_add_hook(
         subprocess.Popen,
-        pre_exec_hook = pysan_hook_subprocess_Popen
+        pre_exec_hook = command_injection.pysan_hook_subprocess_Popen
     )
-
-#pysan_add_hooks()
+    re.compile = pysan_add_hook(
+        re.compile,
+        pre_exec_hook = redos.pysan_hook_re_compile,
+        post_exec_hook = redos.pysan_hook_post_re_compile
+    )
