+++ projects/pysan/pysan/tests/poe/python-ldap-GHSL-2021-117/fuzz_ldap.py
@@ -17,8 +17,12 @@
 import os
 import sys
 import atheris
+
 import pysan
 pysan.pysan_add_hooks()
+
+from pysan import sanlib
+from pysan import redos
 
 import ldap.schema
